Refactor the format reader
